deleted the unused one that had on child, changed all the naming to reflect that we are no longer using onchild
